Finished the activity right before the deadline.
This HTML and CSS activity focuses on my blog. I had fun looking back at the memories I made during my college days. I was planning to add more features to the website, but the deadline is fast approaching. It was a lesson learned for me not to procrastinate on an activity—lol.

Also, I need to work on my CSS skills because I know there are redundant pieces of code I could have avoided if I had reviewed it more carefully.